pyproject.toml
[tool.poetry]
name = "invoice-generator"
version = "0.1.0"
description = "A project to automatically generate invoices, send them via email, store them on Google Drive, and confirm the generation and email delivery via WhatsApp."
authors = ["Emerson-MM-Filho <emerson.dev.machado@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pdfkit = "^1.0.0"
jinja2 = "^3.1.4"
ruff = "^0.6.9"
pre-commit = "^3.8.0"
google-api-python-client = "^2.147.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pydantic-settings = "^2.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
